Phase 0 - Starter Notebook
Downloaded from Piazza - 
Starter Notebook
Sorry for the 3rd thread on this but wanted it to show up on top of list since others are a week old. Professor sent me the correct starter notebook that doesn’t exist in Git and asked me to share it here.  This is what he showed during the lab today.  The one in Git labeled phase 2 is close but this has more content. 

https://www.dropbox.com/s/fows9p5drpra6i6/HCDR-Submission.ipynb?dl=0